v-poc/sdk/aliyun/aliyun.py/requirements.txt
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
crcmod==1.7
cryptography==43.0.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
jmespath==0.10.0
openai==1.47.1
oss2==2.19.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
